addressed several bugs introduced with the lch implementation

+++ golang1.15/test.sh
@@ -17,28 +17,30 @@ cp -R $BASE_DIR/tests/$ID/* $BUILD_DIR
 
 #compile and zip or just zip
 cd $BUILD_DIR
-if [ -z "$COMPILE" ]
-then
-  zip -j main.zip $BUILD_DIR/*
-else
-  CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o exec .
-  echo "openwhisk/action-golang-v1.15" >> exec.env
-  zip main.zip exec exec.env
-  rm exec exec.env
+if [ ! -f "payload.json" ] 
+then 
+  if [ -z "$COMPILE" ]
+  then
+    zip -j main.zip $BUILD_DIR/*
+  else
+    CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o exec .
+    echo "openwhisk/action-golang-v1.15" >> exec.env
+    zip main.zip exec exec.env
+    rm exec exec.env
+  fi
+
+  #generate payload.json
+  echo -n "{\"value\":{\"code\":\"" > payload.json
+  base64 -w 0 main.zip >> payload.json
+  echo -n "\",\"Binary\":true" >> payload.json
+
+  #for non-compiled code set the Main function
+  if [ -z "$COMPILE" ]
+  then
+    echo -n ",\"Main\":\"Main\"" >> payload.json
+  fi
+  echo -n "}}" >> payload.json
 fi
-
-#generate payload.json
-echo -n "{\"value\":{\"code\":\"" > payload.json
-base64 -w 0 main.zip >> payload.json
-echo -n "\",\"Binary\":true" >> payload.json
-
-#for non-compiled code set the Main function
-if [ -z "$COMPILE" ]
-then
-  echo -n ",\"Main\":\"Main\"" >> payload.json
-fi
-echo -n "}}" >> payload.json
-
 #send init command using payload.json
 curl -i -X POST -H "Content-Type: application/json" -d @payload.json http://localhost:8080/init
 sleep 2
